Delete __pycache__/ after unit test is done

+++ python/test/test_application.py
@@ -14,7 +14,8 @@ class TestApplication(unittest.TestCase):
         for i in range(1, 101):
             with open(os.path.join(self.dirname, 'test_file_{i:03}.txt'.format(i = i)), 'w') as f:
                 f.write('')
-        self.pattern = '*.txt'
+        self.pattern  = '*.txt'
+        self.pycaches = glob.glob(os.path.join('.', '**', '__pycache__'))
 
     def test_run_in_dry_run_mode_1(self):
       Application(self.dirname, self.pattern).run()
@@ -29,7 +30,11 @@ class TestApplication(unittest.TestCase):
       self.assertEqual(len(glob.glob(os.path.join(self.dirname, '**', self.pattern), recursive = True)), 0)
 
     def tearDown(self):
-        shutil.rmtree(self.dirname)
+        if os.path.isdir(self.dirname):
+            shutil.rmtree(self.dirname)
+        for pycache in self.pycaches:
+            if os.path.isdir(pycache):
+                shutil.rmtree(pycache)
 
 if __name__ == '__main__':
     unittest.main()
